rename resource
