+++ service/controller/v25/cluster_resource_set.go
@@ -47,9 +47,9 @@ import (
 	"github.com/giantswarm/aws-operator/service/controller/v25/resource/s3bucket"
 	"github.com/giantswarm/aws-operator/service/controller/v25/resource/s3object"
 	"github.com/giantswarm/aws-operator/service/controller/v25/resource/service"
-	"github.com/giantswarm/aws-operator/service/controller/v25/resource/subnet"
 	"github.com/giantswarm/aws-operator/service/controller/v25/resource/tccp"
 	"github.com/giantswarm/aws-operator/service/controller/v25/resource/tccpoutputs"
+	"github.com/giantswarm/aws-operator/service/controller/v25/resource/tccpsubnet"
 	"github.com/giantswarm/aws-operator/service/controller/v25/resource/vpccidr"
 	"github.com/giantswarm/aws-operator/service/controller/v25/resource/workerasgname"
 )
@@ -350,13 +350,13 @@ func NewClusterResourceSet(config ClusterResourceSetConfig) (*controller.Resourc
 		}
 	}
 
-	var subnetResource controller.Resource
+	var tccpSubnetResource controller.Resource
 	{
-		c := subnet.Config{
+		c := tccpsubnet.Config{
 			Logger: config.Logger,
 		}
 
-		subnetResource, err = subnet.New(c)
+		tccpSubnetResource, err = tccpsubnet.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -614,7 +614,7 @@ func NewClusterResourceSet(config ClusterResourceSetConfig) (*controller.Resourc
 		routeTableResource,
 		vpcCIDRResource,
 		tccpOutputsResource,
-		subnetResource,
+		tccpSubnetResource,
 		workerASGNameResource,
 		asgStatusResource,
 		statusResource,
