+++ service/controller/v25/resource/tccpsubnet/error.go
@@ -1,4 +1,4 @@
-package subnet
+package tccpsubnet
 
 import (
 	"github.com/giantswarm/microerror"
