+++ service/controller/v25/resource/tccpsubnet/resource.go
@@ -1,4 +1,4 @@
-package subnet
+package tccpsubnet
 
 import (
 	"context"
